Finished frontend basic flow
Finished the first frontend version now is time to proceed with the backend part
