+++ gptube/components/Header/Header.tsx
@@ -1,7 +1,8 @@
+import Link from "next/link";
 
 const Header:React.FC = () => {
 return <div className="flow-root p-6">
-    <p className="float-left text-5xl text-[#006FFF]">GPTube</p>
+    <Link href="/" className="float-left text-5xl text-[#006FFF]">GPTube</Link>
 </div>;
 };
